subcompositor: split out from compositor
