+++ include/wlr/types/wlr_subcompositor.h
@@ -1,0 +1,37 @@
+/*
+ * This an unstable interface of wlroots. No guarantees are made regarding the
+ * future consistency of this API.
+ */
+#ifndef WLR_USE_UNSTABLE
+#error "Add -DWLR_USE_UNSTABLE to enable unstable wlroots features"
+#endif
+
+#ifndef WLR_TYPES_WLR_SUBCOMPOSITOR_H
+#define WLR_TYPES_WLR_SUBCOMPOSITOR_H
+
+#include <wayland-server-core.h>
+
+struct wlr_surface;
+
+struct wlr_subcompositor {
+	struct wl_global *global;
+
+	struct wl_listener display_destroy;
+
+	struct {
+		struct wl_signal destroy;
+	} events;
+};
+
+bool wlr_surface_is_subsurface(struct wlr_surface *surface);
+
+/**
+ * Get a subsurface from a surface. Can return NULL if the subsurface has been
+ * destroyed.
+ */
+struct wlr_subsurface *wlr_subsurface_from_wlr_surface(
+	struct wlr_surface *surface);
+
+struct wlr_subcompositor *wlr_subcompositor_create(struct wl_display *display);
+
+#endif
